Split the app into custom components accordingly
