+++ src/components/Header.js
@@ -1,0 +1,13 @@
+import React from "react";
+import logo from "../assets/investment-calculator-logo.png";
+
+const Header = () => {
+  return (
+    <header className="header">
+      <img src={logo} alt="logo" />
+      <h1>Investment Calculator</h1>
+    </header>
+  );
+};
+
+export default Header;
